Refactor BookInfo to BookInfoRightSidebar and add documentation
