+++ src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import BookInfo from "../components/BookInfo/BookInfo";
+import BookInfo from "../components/BookInfoRightSidebar/BookInfo";
 import ForYouSection from "../components/ForYouSection/ForYouSection";
 
 export default function Home() {
